+++ docs/changelog.md
@@ -1,6 +1,14 @@
 # TBD
 ### Features
-* The indexer now has a static validator private key, TDOOOOOOOO
+* The indexer now has a static validator private key:
+
+  ```
+  {
+    "account_id": "test.near",
+    "public_key": "ed25519:3Kuyi2DUXdoHgoaNEvCxa1m6G8xqc6Xs7WGajaqLhNmW",
+    "secret_key": "ed25519:2ykcMLiM7vCmsSECcgfmUzihBtNdBv7v2CxNi94sNt4R8ar4xsrMMYvtsSNGQDfSRhNWXEnZvgx2wzS9ViBiS9jW"
+  }
+  ```
 
 ### Fixes
 * The module now correctly uses static files